+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Tap word hints area to show next word.
 - Hide hint box when showing answer.
 - Reset due dates for entire collection.
+- Undo last practice response.
 
 ## 1.3.0 - July 26, 2023